requirements.txt
streamlit>=1.39.0
openai>=1.51.0
python-dotenv>=1.0.1
numpy>=1.26.4
PyPDF2>=3.0.1
